+++ src/main/java/com/dokio/message/response/ReturnsupProductTableJSON.java
@@ -33,7 +33,7 @@ public class ReturnsupProductTableJSON {
     private Boolean     indivisible;                // неделимый товар (нельзя что-то сделать с, например, 0.5 единицами этого товара, только с кратно 1)
     // для печатной версии
     private Integer row_num; // номер строки при выводе печатной версии
-    private Integer nds_value; // сколько % НДС у данного товара
+    private BigDecimal nds_value; // сколько % НДС у данного товара
 
     public Integer getRow_num() {
         return row_num;
@@ -43,14 +43,13 @@ public class ReturnsupProductTableJSON {
         this.row_num = row_num;
     }
 
-    public Integer getNds_value() {
+    public BigDecimal getNds_value() {
         return nds_value;
     }
 
-    public void setNds_value(Integer nds_value) {
+    public void setNds_value(BigDecimal nds_value) {
         this.nds_value = nds_value;
     }
-
 
     public Long getId() {
         return id;
