+++ src/main/java/com/dokio/message/response/additional/InvoiceoutProductTableJSON.java
@@ -46,13 +46,13 @@ public class InvoiceoutProductTableJSON {
     private Boolean     indivisible;         // неделимый товар (нельзя что-то сделать с, например, 0.5 единицами этого товара, только с кратно 1)
     // для печатной версии
     private Integer row_num; // номер строки при выводе печатной версии
-    private Integer nds_value; // сколько % НДС у данного товара
+    private BigDecimal nds_value; // сколько % НДС у данного товара
 
-    public Integer getNds_value() {
+    public BigDecimal getNds_value() {
         return nds_value;
     }
 
-    public void setNds_value(Integer nds_value) {
+    public void setNds_value(BigDecimal nds_value) {
         this.nds_value = nds_value;
     }
