+++ src/main/java/com/dokio/message/response/additional/OrdersupProductTableJSON.java
@@ -39,7 +39,7 @@ public class OrdersupProductTableJSON {
     private Boolean     is_material;                 // материален ли товар
     // для печатной версии
     private Integer row_num;                         // номер строки при выводе печатной версии
-    private Integer nds_value;                       // сколько % НДС у данного товара
+    private BigDecimal nds_value;                       // сколько % НДС у данного товара
 
     public Integer getRow_num() {
         return row_num;
@@ -57,14 +57,13 @@ public class OrdersupProductTableJSON {
         this.row_num = row_num;
     }
 
-    public Integer getNds_value() {
+    public BigDecimal getNds_value() {
         return nds_value;
     }
 
-    public void setNds_value(Integer nds_value) {
+    public void setNds_value(BigDecimal nds_value) {
         this.nds_value = nds_value;
     }
-
 
     public Long getId() {
         return id;
